+++ helpers.go
@@ -1,0 +1,48 @@
+package main
+
+//XXX will likely change post edb/tmint refactor
+
+/*
+import (
+	"bytes"
+	"encoding/json"
+	"fmt"
+	"reflect"
+
+	"github.com/eris-ltd/toadserver/Godeps/_workspace/src/github.com/tendermint/tendermint/wire"
+)
+
+func prettyPrint(o interface{}) (string, error) {
+	var prettyJSON bytes.Buffer
+	err := json.Indent(&prettyJSON, wire.JSONBytes(o), "", "\t")
+	if err != nil {
+		return "", err
+	}
+	return string(prettyJSON.Bytes()), nil
+}
+
+func FieldFromTag(v reflect.Value, field string) (string, error) {
+	iv := v.Interface()
+	st := reflect.TypeOf(iv)
+	for i := 0; i < v.NumField(); i++ {
+		tag := st.Field(i).Tag.Get("json")
+		if tag == field {
+			return st.Field(i).Name, nil
+		}
+	}
+	return "", fmt.Errorf("Invalid field name")
+}
+
+func formatOutput(args []string, i int, o interface{}) (string, error) {
+	if len(args) < i+1 {
+		return prettyPrint(o)
+	}
+	arg0 := args[i]
+	v := reflect.ValueOf(o).Elem()
+	name, err := FieldFromTag(v, arg0)
+	if err != nil {
+		return "", err
+	}
+	f := v.FieldByName(name)
+	return prettyPrint(f.Interface())
+}*/
