+++ main.go
@@ -13,13 +13,10 @@ func main() {
 	//ts makes & signs a nameTx, then posts to a node, which does the broadcasting
 	mux.HandleFunc("/receiveNameTx/", receiveNameTx) //unpack tx, if valid, add to chain
 	mux.HandleFunc("/cacheHash/", cacheHash)         //also if valid, pin hash on all hosts (except one that sent it :~( )
-	mux.HandleFunc("/cacheHashAll/", cacheHashAll)   //also if valid, pin hash on all hosts (except one that sent it :~( )
 
 	mux.HandleFunc("/getfile/", getHandler) // request by name, receive contents
 
 	http.ListenAndServe(":11113", mux)
-
-	//fmt.Println("Initializing toadserver...")
 }
 
 //-------------------------------------------------------
