+++ post_and_get.go
@@ -2,18 +2,18 @@ package main
 
 import (
 	"bytes"
-	"encoding/json"
+	//"encoding/json"
 	"fmt"
 	"io/ioutil"
-	//	"net"
+	//"net"
 	"net/http"
 	"os"
-	"reflect"
+	//"reflect"
 	"strings"
 	"time"
 
 	//"github.com/eris-ltd/mint-client/Godeps/_workspace/src/github.com/tendermint/tendermint/types"
-	"github.com/eris-ltd/toadserver/Godeps/_workspace/src/github.com/tendermint/tendermint/wire"
+	//"github.com/eris-ltd/toadserver/Godeps/_workspace/src/github.com/tendermint/tendermint/wire"
 
 	//cclient "github.com/eris-ltd/toadserver/Godeps/_workspace/src/github.com/tendermint/tendermint/rpc/core_client"
 
@@ -73,121 +73,42 @@ func postHandler(w http.ResponseWriter, r *http.Request) {
 			//return err
 		} else {
 			fmt.Println("Success updating the name registry")
-			//			if err := cacheHashAll(hash); err != nil {
-			//				fmt.Printf("error caching hash: %v\n", err)
-			//			} else {
-			//				fmt.Println("Congratulations, you have successfully added your file to the toadserver")
+			fmt.Println("Caching hash now")
+			if err := cacheHashAll(hash); err != nil {
+				fmt.Printf("error caching hash: %v\n", err)
+			} else {
+				fmt.Println("Congratulations, you have successfully added your file to the toadserver")
+			}
 		}
-		//}
-
 	}
 }
 
-func cacheHashAll(w http.ResponseWriter, r *http.Request) {
-	//func cacheHashAll(hash string) error {
-	str := r.URL.Path[1:]
-	hash := strings.Split(str, "/")[1]
+func cacheHashAll(hash string) error {
 
 	//TODO handle errors to prevent getting here...
 	fmt.Println("Pinning hash to your local IPFS node")
 
-	endpoint := "http://0.0.0.0:11113/" + "cacheHash/" + hash
+	endpoint := fmt.Sprintf("http://0.0.0.0:11113/cacheHash/%s", hash)
 	_, err := http.Post(endpoint, "", nil)
 	if err != nil {
-		w.Write([]byte(fmt.Sprintf("cache post error: %v\n", err)))
+		fmt.Printf("cache post error: %v\n", err)
+		//return err
 	}
 
-	// IPaddrs, _ := getTheNames()
-	IPaddrs := os.Getenv("TOADSERVER_IPFS_NODES") //
+	// IPaddrs, _ := getTheNames() -> use mindy to get ipAddrs
+	IPaddrs := os.Getenv("TOADSERVER_IPFS_NODES")
 	IPs := strings.Split(IPaddrs, ",")
-	//IPaddrs := []string{"toadserver1.interblock.io", "toadserver2.interblock.io", "toadserver3.interblock.io"}
 	for _, ip := range IPs {
-		endpoint := "http://" + ip + ":11113/cacheHash/" + hash
+		endpoint := fmt.Sprintf("http://%s:11113/cacheHash/%s", ip, hash)
 		fmt.Printf("Pinning hash to remote IPFS/toadserver node: %s\n", endpoint)
 		_, err := http.Post(endpoint, "", nil)
 		if err != nil {
 			fmt.Printf("cache post error to endpoint: %s\n%v\n", endpoint, err)
 			continue
+			//TODO return err?
 		}
 	}
-	//return nil
-}
-
-/*func getTheNames() ([]string, error) {
-	mindy := os.Getenv("TOADSERVER_IPFS_NODES")
-	mindyUrl := mindy + ":46657/"
-	//fmt.Printf("Pinning hash to you mindy nodes at:\t%s\n", mindyUrl)
-	c := cclient.NewClient(mindyUrl, "HTTP")
-
-	res, err := c.ListNames()
-	ifExit(err)
-
-	allTheNames, err := formatOutput([]string{}, 1, res)
-	if err != nil {
-		fmt.Printf("error formating output: %v\n")
-	}
-
-	byteNames := []byte(allTheNames)
-
-	rln := struct {
-		BlockHeight int                   `json:"block_height"`
-		Names       []*types.NameRegEntry `json:"names"`
-	}{}
-
-	type NameRegEntry struct {
-		Name    string `json:"name"`    // registered name for the entry
-		Owner   []byte `json:"owner"`   // address that created the entry
-		Data    string `json:"data"`    // data to store under this name
-		Expires int    `json:"expires"` // block at which this entry expires
-	}
-
-	if err := json.Unmarshal(byteNames, &rln); err != nil {
-		return []string{""}, err
-	}
-
-	names := make([]string, len(rln.Names))
-
-	for i, name := range rln.Names {
-		names[i] = name.Name
-	}
-
-	return names, nil
-}
-*/
-
-func prettyPrint(o interface{}) (string, error) {
-	var prettyJSON bytes.Buffer
-	err := json.Indent(&prettyJSON, wire.JSONBytes(o), "", "\t")
-	if err != nil {
-		return "", err
-	}
-	return string(prettyJSON.Bytes()), nil
-}
-
-func FieldFromTag(v reflect.Value, field string) (string, error) {
-	iv := v.Interface()
-	st := reflect.TypeOf(iv)
-	for i := 0; i < v.NumField(); i++ {
-		tag := st.Field(i).Tag.Get("json")
-		if tag == field {
-			return st.Field(i).Name, nil
-		}
-	}
-	return "", fmt.Errorf("Invalid field name")
-}
-
-func formatOutput(args []string, i int, o interface{}) (string, error) {
-	if len(args) < i+1 {
-		return prettyPrint(o)
-	}
-	arg0 := args[i]
-	v := reflect.ValueOf(o).Elem()
-	name, err := FieldFromTag(v, arg0)
-	if err != nil {
-		return "", err
-	}
-	f := v.FieldByName(name)
-	return prettyPrint(f.Interface())
+	return nil
 }
 
 func getHandler(w http.ResponseWriter, r *http.Request) {
@@ -244,7 +165,6 @@ func getHandler(w http.ResponseWriter, r *http.Request) {
 }
 
 //XXX this endpoint should require authentication
-//if name reg update is legit, make post to ipfs hosts (using mindy!) w/ hash in URL;
 func cacheHash(w http.ResponseWriter, r *http.Request) {
 	str := r.URL.Path[1:]
 	hash := strings.Split(str, "/")[1]
@@ -257,3 +177,49 @@ func cacheHash(w http.ResponseWriter, r *http.Request) {
 		w.Write([]byte(fmt.Sprintf("Caching succesful:\t%s\n", pinned)))
 	}
 }
+
+//XXX for getting IPaddrs that will do all the caching
+// (rather than pass in TOADSERVER_IPFS_NODES as a csv string of addrs)
+// need to think of the name spacing carefullly
+// this'll block on the tendermint refactor
+/*func getTheNames() ([]string, error) {
+	mindy := os.Getenv("TOADSERVER_IPFS_NODES")
+	mindyUrl := mindy + ":46657/"
+	//fmt.Printf("Pinning hash to you mindy nodes at:\t%s\n", mindyUrl)
+	c := cclient.NewClient(mindyUrl, "HTTP")
+
+	res, err := c.ListNames()
+	ifExit(err)
+
+	allTheNames, err := formatOutput([]string{}, 1, res)
+	if err != nil {
+		fmt.Printf("error formating output: %v\n")
+	}
+
+	byteNames := []byte(allTheNames)
+
+	rln := struct {
+		BlockHeight int                   `json:"block_height"`
+		Names       []*types.NameRegEntry `json:"names"`
+	}{}
+
+	type NameRegEntry struct {
+		Name    string `json:"name"`    // registered name for the entry
+		Owner   []byte `json:"owner"`   // address that created the entry
+		Data    string `json:"data"`    // data to store under this name
+		Expires int    `json:"expires"` // block at which this entry expires
+	}
+
+	if err := json.Unmarshal(byteNames, &rln); err != nil {
+		return []string{""}, err
+	}
+
+	names := make([]string, len(rln.Names))
+
+	for i, name := range rln.Names {
+		names[i] = name.Name
+	}
+
+	return names, nil
+}
+*/
